move layers to a sub-module
