+++ .gitmodules
@@ -1,0 +1,3 @@
+[submodule "layers"]
+	path = layers
+	url = git@github.com:mead-ml/mead-layers.git
