--- python/baseline/mime_type.py
@@ -1,5 +1,0 @@
-import eight_mile.mime_type
-from eight_mile.mime_type import *
-
-__all__ = []
-__all__.extend(eight_mile.mime_type.__all__)
