--- python/baseline/w2v.py
@@ -1,5 +1,0 @@
-import eight_mile.w2v
-from eight_mile.w2v import *
-
-__all__ = []
-__all__.extend(eight_mile.w2v.__all__)
